Made Partition into its own class & refactored stuff to make that work
