+++ src/test/java/com/datastax/cdm/cql/statement/TargetUpsertRunDetailsStatementTest.java
@@ -32,8 +32,8 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 
 import com.datastax.cdm.cql.CommonMocks;
+import com.datastax.cdm.job.Partition;
 import com.datastax.cdm.job.RunNotStartedException;
-import com.datastax.cdm.job.SplitPartitions;
 import com.datastax.oss.driver.api.core.CqlSession;
 import com.datastax.oss.driver.api.core.cql.BoundStatement;
 import com.datastax.oss.driver.api.core.cql.PreparedStatement;
@@ -98,8 +98,7 @@ public class TargetUpsertRunDetailsStatementTest extends CommonMocks {
         when(mockIterator.next()).thenReturn(row3);
 
         targetUpsertRunDetailsStatement = new TargetUpsertRunDetailsStatement(cqlSession, "ks.table1");
-        Collection<SplitPartitions.Partition> parts = targetUpsertRunDetailsStatement.getPartitionsByStatus(123l,
-                "RUNNING");
+        Collection<Partition> parts = targetUpsertRunDetailsStatement.getPartitionsByStatus(123l, "RUNNING");
 
         // This test is incorrect, but needs to be troubleshot & fixed. The actual code works, but the test does not
         assertEquals(0, parts.size());
