+++ src/test/java/com/datastax/cdm/job/SplitPartitionsTest.java
@@ -32,8 +32,8 @@ public class SplitPartitionsTest {
 
     @Test
     void getRandomSubPartitionsTest() {
-        List<SplitPartitions.Partition> partitions = SplitPartitions.getRandomSubPartitions(10, BigInteger.ONE,
-                BigInteger.valueOf(100), 100);
+        List<Partition> partitions = SplitPartitions.getRandomSubPartitions(10, BigInteger.ONE, BigInteger.valueOf(100),
+                100);
         assertEquals(10, partitions.size());
         partitions.forEach(p -> {
             assertEquals(9, p.getMax().longValue() - p.getMin().longValue());
@@ -42,8 +42,8 @@ public class SplitPartitionsTest {
 
     @Test
     void getRandomSubPartitionsTestOver100() {
-        List<SplitPartitions.Partition> partitions = SplitPartitions.getRandomSubPartitions(8, BigInteger.ONE,
-                BigInteger.valueOf(44), 200);
+        List<Partition> partitions = SplitPartitions.getRandomSubPartitions(8, BigInteger.ONE, BigInteger.valueOf(44),
+                200);
         assertEquals(8, partitions.size());
     }
